[FIX] 검증 추가, ApiResponse 임포트, AuthController -> MemberController 변경
